refactor: isolate keyboard class
