+++ codejam-virtual-keyboard/src/data/keyboard.json
@@ -433,7 +433,8 @@
       "type": "system",
       "code": "ShiftRight",
       "value": "Shift",
-      "className": "key-button--shift-right"
+      "className": "key-button--shift-right",
+      "action": "shift"
     }
   ],
   [
@@ -441,8 +442,7 @@
       "type": "system",
       "code": "ControlLeft",
       "value": "Ctrl",
-      "className": "key-button--ctrl-left",
-      "action": "ctrl"
+      "className": "key-button--ctrl-left"
     },
     {
       "type": "system",
